requirement.txt
numpy==1.21.0
Pillow==8.2.0
matplotlib==3.4.3
scikit-image==0.18.3
opencv-python==4.5.2
